fnc_arcor_orders_process/requirements.txt
functions-framework==3.*
google-cloud>=0.34.0
google-cloud-storage>=2.7.0
google-cloud-bigquery>=2.7.0
python-dateutil